Move legacy pages to `/pages/legacy`
move new pages out of `/pages/new` to `/pages`
