+++ evac_app/lib/pages/legacy/wait_screen.dart
