+++ evac_app/lib/presenters/basic_drill_presenter.dart
@@ -3,13 +3,13 @@ import 'dart:convert';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:evac_app/export/result_exporter.dart';
 import 'package:evac_app/models/drill_result.dart';
-import 'package:evac_app/pages/confirm_drill.dart';
-import 'package:evac_app/pages/during_drill.dart';
-import 'package:evac_app/pages/invite_code_page.dart';
-import 'package:evac_app/pages/landing_page.dart';
-import 'package:evac_app/pages/post_drill_survey.dart';
-import 'package:evac_app/pages/pre_drill_survey.dart';
-import 'package:evac_app/pages/wait_screen.dart';
+import 'package:evac_app/pages/legacy/confirm_drill.dart';
+import 'package:evac_app/pages/legacy/during_drill.dart';
+import 'package:evac_app/pages/legacy/invite_code_page.dart';
+import 'package:evac_app/pages/legacy/landing_page.dart';
+import 'package:evac_app/pages/legacy/post_drill_survey.dart';
+import 'package:evac_app/pages/legacy/pre_drill_survey.dart';
+import 'package:evac_app/pages/legacy/wait_screen.dart';
 import 'package:evac_app/models/drill_event.dart';
 import 'package:flutter/material.dart';
 import 'package:survey_kit/survey_kit.dart';
